+++ DiceChecker/discreteFourierTransformTest.h
@@ -1,8 +1,8 @@
 //
 // Creator:    http://www.dicelocksecurity.com
-// Version:    vers.4.0.0.1
+// Version:    vers.5.0.0.1
 //
-// Copyright © 2008-2010 DiceLock Security, LLC. All rigths reserved.
+// Copyright © 2008-2011 DiceLock Security, LLC. All rights reserved.
 //
 //                               DISCLAIMER
 //
@@ -44,12 +44,12 @@ namespace DiceLockSecurity {
 
 		protected:
 
-			// Random Test Class enumerator name
+			/// Random Test Class enumerator name
 			static const RandomTests	test;
-			// Random Test Class minimum stream length
-			static const unsigned int	minimumLength;
+			/// Random Test Class minimum stream length
+			static const unsigned long int	minimumLength;
 
-			static const int	ntryh[4];
+			static const signed long int ntryh[4];
 			static const double	tpi;
 			static const double	hsqt2;
 
@@ -58,48 +58,48 @@ namespace DiceLockSecurity {
 			double      expectedPeaks;
 			double      normalizedDifference;
 
-			void drfti1(int, double *, int *);
-			void dradf2(int, int, double *, double *, double *);
-			void dradf4(int, int, double *, double *, double *, double *, double *);
-			void drftf1(int, double *, double *, double *, int *);
-			void dradfg(int, int, int, int, double *, double *, double *, double *, double *, double *);
+			void drfti1(signed long int, double *, signed long int *);
+			void dradf2(signed long int, signed long int, double *, double *, double *);
+			void dradf4(signed long int, signed long int, double *, double *, double *, double *, double *);
+			void drftf1(signed long int, double *, double *, double *, signed long int *);
+			void dradfg(signed long int, signed long int, signed long int, signed long int, double *, double *, double *, double *, double *, double *);
 
 		public:
 
-			// Constructor, default 
+			/// Constructor, default 
 			CLASS_DECLSPEC DiscreteFourierTransformTest();
 
-			// Constructor with a MathematicalFunctions object instantiated 
+			/// Constructor with a MathematicalFunctions object instantiated 
 			CLASS_DECLSPEC DiscreteFourierTransformTest(MathematicalFunctions*);
 
-			// Destructor
+			/// Destructor
 			CLASS_DECLSPEC ~DiscreteFourierTransformTest();
 
-			// Gets the BaseRandomTest random state of the last executed BaseCryptoRandomStream
+			/// Gets the BaseRandomTest random state of the last executed BaseCryptoRandomStream
 			CLASS_DECLSPEC bool IsRandom(void);
 
-			// Tests randomness of the BaseCryptoRandomStream and returns the random value
+			/// Tests randomness of the BaseCryptoRandomStream and returns the random value
 			CLASS_DECLSPEC bool IsRandom(BaseCryptoRandomStream*);
 
-			// Initializes the object
+			/// Initializes the object
 			CLASS_DECLSPEC void Initialize(void);
 
-			// Gets the type of the object
+			/// Gets the type of the object
 			CLASS_DECLSPEC RandomTests GetType(void);
 
-			// Gets the minimum random stream length
-			CLASS_DECLSPEC unsigned int GetMinimumLength(void);
+			/// Gets the minimum random stream length
+			CLASS_DECLSPEC unsigned long int GetMinimumLength(void);
 
-			// Gets the "percentile" result
+			/// Gets the "percentile" result
 			CLASS_DECLSPEC double GetPercentile(void);
 
-			// Gets the "observedPeaks" result
+			/// Gets the "observedPeaks" result
 			CLASS_DECLSPEC double GetObservedPeaks(void);
 
-			// Gets the "expectedPeaks" result
+			/// Gets the "expectedPeaks" result
 			CLASS_DECLSPEC double GetExpectedPeaks(void);
 
-			// Gets the "normalizedDifference" result
+			/// Gets the "normalizedDifference" result
 			CLASS_DECLSPEC double GetNormalizedDifference (void);
 	};
   }
